pynalyzer/requirements.txt
gitdb2==2.0.3
GitPython==2.1.8
nltk==3.2.5
six==1.11.0
smmap2==2.0.3
